exclude test files from gem
Files referenced via test_files in the gemspec get included in the gem on build time. As there is a test application in the spec directory this will lead to the gem expanding to 600+MB.

This behaviour has been discussed extensively (https://github.com/rubygems/rubygems/issues/735) but a fix is not in sight.